requirements.txt
ipykernel
pandas
openai
langchain

nbformat
google-auth
google-api-python-client
google-cloud-storage